requirements.txt
Django==5.0.2
python-dotenv==1.0.1
openai==1.12.0
Pillow==10.2.0 